refactor: features/에서 pages/ 레이어로 페이지 컴포넌트 이동
- FSD 아키텍처를 따라 라우트 단위 페이지 컴포넌트 분리
- cardComplete와 cardRegistration을 features에서 pages로 이동
- 페이지 레이아웃과 기능 관련 컴포넌트의 관심사 분리 개선
