+++ src/pages/cardRegistration/cardRegistration.css
